+++ easypost/services/carrier_metadata_service.py
@@ -4,7 +4,6 @@ from typing import (
     List,
     Optional,
 )
-from warnings import warn
 
 from easypost.easypost_object import convert_to_easypost_object
 from easypost.requestor import (
@@ -14,22 +13,16 @@ from easypost.requestor import (
 from easypost.services.base_service import BaseService
 
 
-class BetaCarrierMetadataService(BaseService):
+class CarrierMetadataService(BaseService):
     def __init__(self, client):
         self._client = client
 
-    def retrieve_carrier_metadata(
+    def retrieve(
         self,
         carriers: Optional[List[str]] = None,
         types: Optional[List[str]] = None,
     ) -> List[Dict[str, Any]]:
-        """Get metadata for all carriers on the EasyPost platform."""
-        warn(
-            'This method is deprecated, use the "retrieve" function of "carrier_metadata" on the client instead.',
-            DeprecationWarning,
-            stacklevel=2,
-        )
-
+        """Get metadata for all carriers on the EasyPost platform or specify optional filters."""
         params = {
             "carriers": ",".join(carriers) if carriers else None,
             "types": ",".join(types) if types else None,
@@ -37,9 +30,8 @@ class BetaCarrierMetadataService(BaseService):
 
         response = Requestor(self._client).request(
             method=RequestMethod.GET,
-            url="/metadata",
+            url="/metadata/carriers",
             params=params,
-            beta=True,
         )
 
         return convert_to_easypost_object(response=response.get("carriers", []))